q1f-trading-bot/requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
ta-lib>=0.4.28
python-telegram-bot>=20.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
